changed DeliveryCost type to decimal(19, 4)

+++ src/Core/Logistics.Application.Tenant/Commands/CreateLoad/CreateLoadCommand.cs
@@ -9,7 +9,7 @@ public class CreateLoadCommand : Request<ResponseResult>
     public string? DestinationAddress { get; set; }
     public double DestinationAddressLat { get; set; }
     public double DestinationAddressLong { get; set; }
-    public double DeliveryCost { get; set; }
+    public decimal DeliveryCost { get; set; }
     public double Distance { get; set; }
     public string? AssignedDispatcherId { get; set; }
     public string? AssignedTruckId { get; set; }
