+++ src/Core/Logistics.Application.Tenant/Commands/UpdateLoad/UpdateLoadCommand.cs
@@ -12,7 +12,7 @@ public class UpdateLoadCommand : Request<ResponseResult>
     public string? DestinationAddress { get; set; }
     public double? DestinationAddressLat { get; set; }
     public double? DestinationAddressLong { get; set; }
-    public double? DeliveryCost { get; set; }
+    public decimal? DeliveryCost { get; set; }
     public double? Distance { get; set; }
     public string? AssignedDispatcherId { get; set; }
     public string? AssignedTruckId { get; set; }
