+++ src/Shared/Logistics.Models/LoadDto.cs
@@ -11,7 +11,7 @@ public class LoadDto
     public string? DestinationAddress { get; set; }
     public double? DestinationAddressLat { get; set; }
     public double? DestinationAddressLong { get; set; }
-    public double DeliveryCost { get; set; }
+    public decimal DeliveryCost { get; set; }
     public double Distance { get; set; }
     public DateTime DispatchedDate { get; set; }
     public DateTime? PickUpDate { get; set; }
