+++ src/Shared/Logistics.Models/MonthlyGrossDto.cs
@@ -19,7 +19,7 @@ public record MonthlyGrossDto : IGrossChart
     }
 
     public DateTime Date { get; set; }
-    public double DriverShare { get; set; }
-    public double Gross { get; set; }
+    public decimal DriverShare { get; set; }
+    public decimal Gross { get; set; }
     public double Distance { get; set; }
 }
